requirements.txt
# Recipe RAG 依赖清单

# LlamaIndex 核心框架
llama-index-core
llama-index-vector-stores-qdrant
llama-index-embeddings-huggingface
llama-index-llms-dashscope

# 向量数据库
qdrant-client

# Embedding 模型运行时
sentence-transformers

# Web UI
gradio

# BM25 混合检索
rank-bm25
jieba

# Web 后端（FastAPI）
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
aiosqlite
python-multipart

# 工具库
numpy
python-dotenv
opencc-python-reimplemented
pytest
pytest-asyncio
httpx
